refactor: constants 파일 및 폴더명 수정
