+++ src/utils/common.ts
@@ -1,4 +1,5 @@
 /* 랜덤 숫자 생성 */
 export function getUniqueNumber() {
+  // TODO: 메뉴리스트 스토어 구현 후 중복검사 코드 추가 예정
   return Math.floor(Math.random() * 100);
 }
