+++ src/utils/domController.ts
@@ -18,23 +18,6 @@ export const addElement = (options: ElCreationOptions) => {
   return $el;
 };
 
-type ElAttributesOptions = {
-  TYPE: string;
-  ID: string;
-  CLASS_NAME: string;
-  TEXT?: string;
-};
-
-export const getElCreationOptions = (constants: ElAttributesOptions) => {
-  const { TYPE, ID, CLASS_NAME, TEXT } = constants;
-  return {
-    elName: TYPE,
-    id: ID,
-    className: CLASS_NAME,
-    text: TEXT,
-  };
-};
-
 export const bindEvent = (
   $el: HTMLElement,
   eventName: string,
